+++ CHANGELOG.md
@@ -3,14 +3,16 @@
 All notable changes to this project will be documented in this file.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [0.3.2] In Progress
+## [0.3.2] 22-07-2025
 
 ### Added 
 - SettingsProvider
 - FlatIcon to markdown attribution fixer
 
-### Fixed
+### Changed
+- Links to Essential Packages is in Project Settings now. Choose between Dev or Stable.
 
+### Fixed
 - Path in demo video
 - Unity Version "6000.0"